refactor: rename arquives to camel case name pattern
